# docs/cms-api/swagger.yaml
basePath: /
definitions:
...
  github_com_hadialqattan_mediacms_internal_shared_domain.Program:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      description:
        type: string
      duration_ms:
        type: integer
      id:
        type: string
      language:
        $ref: '#/definitions/github_com_hadialqattan_mediacms_internal_shared_domain.ProgramLanguage'
      published_at:
        type: string
      published_by:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        $ref: '#/definitions/github_com_hadialqattan_mediacms_internal_shared_domain.ProgramType'
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  github_com_hadialqattan_mediacms_internal_shared_domain.ProgramLanguage:
    enum:
    - ar
    - en
    type: string
    x-enum-varnames:
    - LanguageAr
    - LanguageEn
  github_com_hadialqattan_mediacms_internal_shared_domain.ProgramType:
    enum:
    - podcast
    - documentary
    type: string
    x-enum-varnames:
    - ProgramTypePodcast
    - ProgramTypeDocumentary
...
  internal_cms_handler.CreateProgramRequest:
    properties:
      description:
        example: A great podcast about technology
        type: string
      duration_ms:
        example: 3600000
        type: integer
      language:
        allOf:
        - $ref: '#/definitions/github_com_hadialqattan_mediacms_internal_shared_domain.ProgramLanguage'
        enum:
        - ar
        - en
        example: ar
      slug:
        example: my-podcast
        type: string
      tags:
        example:
        - news
        - sports
        items:
          type: string
        type: array
      title:
        example: My Podcast
        type: string
      type:
        allOf:
        - $ref: '#/definitions/github_com_hadialqattan_mediacms_internal_shared_domain.ProgramType'
        enum:
        - podcast
        - documentary
        example: podcast
    type: object
  internal_cms_handler.ListProgramsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_hadialqattan_mediacms_internal_shared_domain.Program'
        type: array
      meta:
        $ref: '#/definitions/internal_cms_handler.PaginationMeta'
    type: object
...
  internal_cms_handler.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
    type: object
...
  internal_cms_handler.UpdateProgramRequest:
    properties:
      description:
        type: string
      duration_ms:
        type: integer
      language:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Content Management System API for MediaCMS
  title: MediaCMS API
  version: "1.0"
paths:
...
  /api/v1/programs:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all programs
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_cms_handler.ListProgramsResponse'
        "500":
          description: Failed to list programs
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List programs
      tags:
      - programs
    post:
      consumes:
      - application/json
      description: Create a new podcast or documentary program
      parameters:
      - description: Program data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_cms_handler.CreateProgramRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_hadialqattan_mediacms_internal_shared_domain.Program'
        "400":
          description: Invalid type or language
          schema:
            type: string
        "409":
          description: Program with this slug already exists
          schema:
            type: string
        "500":
          description: Failed to create program
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create program
      tags:
      - programs
  /api/v1/programs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a program by ID (soft delete)
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "500":
          description: Failed to delete program
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete program
      tags:
      - programs
    get:
      consumes:
      - application/json
      description: Get a specific program by its ID
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hadialqattan_mediacms_internal_shared_domain.Program'
        "404":
          description: Program not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get program
      tags:
      - programs
    put:
      consumes:
      - application/json
      description: Update an existing program by ID
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: string
      - description: Program update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_cms_handler.UpdateProgramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hadialqattan_mediacms_internal_shared_domain.Program'
        "400":
          description: Invalid type or language
          schema:
            type: string
        "404":
          description: Program not found
          schema:
            type: string
        "500":
          description: Failed to update program
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update program
      tags:
      - programs
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"